Add styling to analytical statement
