+++ src/typings/analysisModule.tsx
@@ -36,7 +36,7 @@ export interface AnalysisPillarFormItem {
 }
 
 export interface PillarAnalysisElement extends AnalysisPillars {
-    analysisName: string;
+    analysisTitle: string;
     title: string;
     filters?: PillarFilterItem[];
 }
